Rename page components to start with "Page"
