+++ src/components/productdetail/PageProductDetail.vue
